@@ -1,6 +1,6 @@
 {
   "name": "@kintone/rest-api-client-demo",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "private": true,
   "author": {
     "name": "Cybozu, Inc.",
@@ -13,6 +13,8 @@
     "directory": "examples/rest-api-client-demo"
   },
   "scripts": {
+    "prebuild": "rimraf dist",
+    "build": "tsc --build",
     "fix:lint": "npm run lint:eslint -- --fix",
     "fix:md": "prettier --write README.md",
     "fix": "run-p fix:*",
@@ -22,7 +24,7 @@
     "lint": "run-p -l lint:*",
     "deploy": "rimraf scripts/dist && run-s webpack upload",
     "upload": "kintone-customize-uploader customize-manifest.json",
-    "run-script": "ts-node -r tsconfig-paths/register src/run-node.ts",
+    "run-script": "node dist/run-node.js",
     "webpack": "webpack"
   },
   "bugs": {
@@ -32,17 +34,15 @@
     "node": ">=10.0.0"
   },
   "dependencies": {
-    "@kintone/profile-loader": "^1.1.1",
-    "@kintone/rest-api-client": "^1.9.0",
+    "@kintone/profile-loader": "^1.1.2",
+    "@kintone/rest-api-client": "^1.10.0",
     "yargs": "^16.2.0"
   },
   "devDependencies": {
-    "@kintone/customize-uploader": "^3.1.4",
+    "@kintone/customize-uploader": "^3.1.5",
     "@types/yargs": "^16.0.0",
     "ts-loader": "^8.0.17",
-    "ts-node": "^9.1.1",
-    "tsconfig-paths": "^3.9.0",
-    "webpack": "^5.21.2",
+    "webpack": "^5.23.0",
     "webpack-cli": "4.5.0"
   }
 }
