@@ -8,15 +8,17 @@
     "examples/*"
   ],
   "scripts": {
-    "build": "lerna run build --stream",
-    "prelint": "yarn build",
-    "lint": "lerna run lint --stream",
+    "prebuild": "lerna run prebuild --parallel",
+    "build": "tsc --build packages --force",
+    "postbuild": "lerna run postbuild --parallel",
+    "lint": "lerna run lint --parallel",
     "preinitial:release": "yarn prerelease",
     "initial:release": "lerna publish from-package --conventional-commits",
-    "prerelease": "lerna run prerelease --stream",
+    "prerelease": "npm-run-all -s build -p lint test",
     "release": "lerna publish --conventional-commits",
-    "pretest": "yarn build && jest --rootDir __tests__",
-    "test": "lerna run test --stream",
+    "pretest": "jest --rootDir __tests__",
+    "start": "yarn build --watch",
+    "test": "lerna run test --parallel",
     "pretest:ci": "yarn pretest",
     "test:ci": "lerna run test:ci --stream"
   },
@@ -30,13 +32,14 @@
   "author": "kintone",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.12.16",
-    "@babel/preset-env": "^7.12.16",
-    "@babel/preset-typescript": "^7.12.16",
+    "@babel/core": "^7.12.17",
+    "@babel/preset-env": "^7.12.17",
+    "@babel/preset-typescript": "^7.12.17",
     "@cybozu/eslint-config": "^11.0.4",
     "@types/jest": "^26.0.20",
-    "@types/node": "^13.13.42",
-    "eslint": "^7.19.0",
+    "@types/node": "^13.13.45",
+    "comment-json": "^4.1.0",
+    "eslint": "^7.20.0",
     "jest": "^26.6.3",
     "lerna": "^3.22.1",
     "npm-run-all": "^4.1.5",
