@@ -1,21 +1,20 @@
 {
   "name": "@kintone/dts-gen",
-  "version": "3.1.3",
+  "version": "3.1.4",
   "description": "Types for kintone js api and Types generating tools",
   "main": "dist/index.js",
   "scripts": {
-    "build": "bin/npm-build.sh",
-    "tsc": "tsc",
-    "tsc:check": "tsc --noEmit",
-    "clean": "rm -rfv dist",
+    "prebuild": "yarn clean",
+    "build": "tsc --build --force",
+    "postbuild": "bin/npm-build.sh && yarn build:integration",
+    "clean": "rimraf dist",
     "lint:fix": "eslint --fix src/**/*.ts",
     "lint": "eslint 'src/**/*.ts'",
     "demo": "node dist/index.js --demo --type-name DemoFields -o demo-fields.d.ts",
     "generate": "node dist/index.js --host https://****.cybozu.com --username *** --password *** --app-id ***",
     "help": "node dist/index.js --help",
-    "prerelease": "run-s lint test clean build",
     "test": "jest",
-    "test:ci": "jest",
+    "test:ci": "jest --runInBand",
     "prettier:check": "prettier --parser typescript --check \"src/**/*.{ts,tsx}\" ./kintone.d.ts",
     "fmt": "prettier --parser typescript --write \"src/**/*.{ts,tsx}\" ./kintone.d.ts",
     "build:integration": "webpack --mode development --config webpack.config.js"
@@ -30,15 +29,15 @@
   "license": "MIT",
   "dependencies": {
     "axios": "^0.21.0",
-    "commander": "^7.0.0",
-    "form-data": "^3.0.0",
-    "lodash": "^4.17.20",
+    "commander": "^7.1.0",
+    "form-data": "^4.0.0",
+    "lodash": "^4.17.21",
     "prettier": "^2.2.1"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.168",
     "ts-loader": "^8.0.17",
-    "webpack": "^5.21.2",
+    "webpack": "^5.23.0",
     "webpack-cli": "^4.5.0"
   },
   "bin": {
