@@ -1,21 +1,19 @@
 {
   "name": "@kintone/webpack-plugin-kintone-plugin",
-  "version": "4.0.12",
+  "version": "4.0.13",
   "description": "A webpack plugin to generate a plugin zip",
   "main": "dist/index.js",
   "engines": {
     "node": ">=10"
   },
   "scripts": {
-    "start": "tsc -w",
-    "build": "run-s clean build:tsc",
-    "build:tsc": "tsc",
+    "start": "yarn build --watch",
+    "prebuild": "yarn clean",
+    "build": "tsc --build --force",
     "clean": "rimraf dist",
     "lint": "eslint 'src/**/*.ts'",
-    "prerelease": "run-p build lint test",
-    "pretest": "run-s build",
     "test": "jest",
-    "test:ci": "run-s test"
+    "test:ci": "test --runInBand"
   },
   "keywords": [
     "webpack",
@@ -44,19 +42,19 @@
     "@types/lodash": "^4.14.168",
     "@types/mkdirp": "^1.0.1",
     "@types/rimraf": "^3.0.0",
-    "adm-zip": "^0.5.2",
+    "adm-zip": "^0.5.3",
     "rimraf": "^3.0.2",
     "ts-node": "^9.1.1",
-    "webpack": "^5.21.2",
+    "webpack": "^5.23.0",
     "webpack-cli": "4.5.0"
   },
   "peerDependencies": {
     "webpack": "^4.0.0 || ^5.0.0"
   },
   "dependencies": {
-    "@kintone/plugin-packer": "^3.0.13",
+    "@kintone/plugin-packer": "^3.0.14",
     "chokidar": "^3.5.1",
-    "lodash": "^4.17.20",
+    "lodash": "^4.17.21",
     "mkdirp": "^1.0.4"
   }
 }
