@@ -1,11 +1,7 @@
-"use strict";
-
-const path = require("path");
-const fs = require("fs");
-
-const { readZipContentsNames } = require("./helper/zip");
-
-const createContentsZip = require("../src/create-contents-zip");
+import path from "path";
+import fs from "fs";
+import { readZipContentsNames } from "./helper/zip";
+import { createContentsZip } from "../src/create-contents-zip";
 
 const fixturesDir = path.join(__dirname, "fixtures");
 const pluginDir = path.join(fixturesDir, "sample-plugin", "plugin-dir");
@@ -15,7 +11,7 @@ describe("create-contents-zip", () => {
     const manifestJSONPath = path.join(pluginDir, "manifest.json");
     const manifest = JSON.parse(fs.readFileSync(manifestJSONPath, "utf-8"));
     createContentsZip(pluginDir, manifest).then((buffer) => {
-      readZipContentsNames(buffer).then((files) => {
+      readZipContentsNames(buffer as Buffer).then((files) => {
         expect(files).toStrictEqual(["manifest.json", "image/icon.png"]);
         expect(buffer).toBeInstanceOf(Buffer);
         done();
