@@ -1,6 +1,5 @@
 import { promises as fs } from "fs";
 import path from "path";
-import PQueue from "p-queue";
 
 import { KintoneRestAPIClient } from "@kintone/rest-api-client";
 import { AppID, Record } from "@kintone/rest-api-client/lib/client/types";
@@ -40,24 +39,10 @@ export async function exportRecords(
 
   // TODO: extract attachment fields first
 
-  if (!attachmentDir) {
-    return records;
+  if (attachmentDir) {
+    await downloadAttachments(apiClient, records, attachmentDir);
   }
-  // download attachments if exists
-  const fetchFiles = async (record: Record) => {
-    const fileInfos = getFileInfos(record);
-    for (const fileInfo of fileInfos) {
-      await downloadAttachments(apiClient, record, attachmentDir, fileInfo);
-    }
-  };
-  const queue = new PQueue({ concurrency: 5 });
-  await queue.addAll(
-    records.map((record: Record) => {
-      return () => {
-        return fetchFiles(record);
-      };
-    })
-  );
+
   return records;
 }
 
@@ -77,14 +62,19 @@ const getFileInfos = (record: Record) => {
 
 const downloadAttachments = async (
   apiClient: KintoneRestAPIClient,
-  record: Record,
-  attachmentDir: string,
-  fileInfo: FileInfo
+  records: Record[],
+  attachmentDir: string
 ) => {
-  const { fileKey, name } = fileInfo;
-  const file = await apiClient.file.downloadFile({ fileKey });
-  const recordId = record.$id.value as string;
-  const dir = path.resolve(attachmentDir, recordId);
-  await fs.mkdir(dir, { recursive: true });
-  await fs.writeFile(path.resolve(dir, name), Buffer.from(file));
+  for (const record of records) {
+    const fileInfos = getFileInfos(record);
+    for (const fileInfo of fileInfos) {
+      const { fileKey, name } = fileInfo;
+      const file = await apiClient.file.downloadFile({ fileKey });
+
+      const recordId = record.$id.value as string;
+      const dir = path.resolve(attachmentDir, recordId);
+      await fs.mkdir(dir, { recursive: true });
+      await fs.writeFile(path.resolve(dir, name), Buffer.from(file));
+    }
+  }
 };
