@@ -1,27 +1,27 @@
-"use strict";
+import path from "path";
+import fs from "fs";
+import { promisify } from "util";
+import os from "os";
+import * as chokidar from "chokidar";
+import mkdirp from "mkdirp";
+import _debug from "debug";
+import validate from "@kintone/plugin-manifest-validator";
+import packer from ".";
+import console from "./console";
+import { generateErrorMessages } from "./gen-error-msg";
+import { createContentsZip } from "./create-contents-zip";
 
-const path = require("path");
-const fs = require("fs");
-const os = require("os");
-const chokidar = require("chokidar");
-const denodeify = require("denodeify");
+const debug = _debug("cli");
+const writeFile = promisify(fs.writeFile);
 
-const writeFile = denodeify(fs.writeFile);
-const mkdirp = require("mkdirp");
-const debug = require("debug")("cli");
-const validate = require("@kintone/plugin-manifest-validator");
+type Options = Partial<{
+  ppk: string;
+  out: string;
+  watch: boolean;
+  packerMock_: typeof packer;
+}>;
 
-const packer = require("./");
-const console = require("./console");
-const generateErrorMessages = require("./gen-error-msg");
-const createContentsZip = require("./create-contents-zip");
-
-/**
- * @param {string} pluginDir path to plugin directory.
- * @param {Object=} options {ppk: string, out: string}.
- * @return {!Promise<string>} The resolved value is a path to the output plugin zip file.
- */
-function cli(pluginDir, options_) {
+export = function cli(pluginDir: string, options_?: Options) {
   const options = options_ || {};
   const packerLocal = options.packerMock_ ? options.packerMock_ : packer;
 
@@ -53,7 +53,7 @@ function cli(pluginDir, options_) {
 
       // 4. generate new ppk if not specified
       const ppkFile = options.ppk;
-      let privateKey;
+      let privateKey: string;
       if (ppkFile) {
         debug(`loading an existing key: ${ppkFile}`);
         privateKey = fs.readFileSync(ppkFile, "utf8");
@@ -107,15 +107,9 @@ function cli(pluginDir, options_) {
       console.error("Failed:", error.message);
       return Promise.reject(error);
     });
-}
+};
 
-module.exports = cli;
-
-/**
- * @param {!Object} manifest
- * @param {string} pluginDir
- */
-function throwIfInvalidManifest(manifest, pluginDir) {
+function throwIfInvalidManifest(manifest: any, pluginDir: string) {
   const result = validate(manifest, {
     relativePath: validateRelativePath(pluginDir),
     maxFileSize: validateMaxFileSize(pluginDir),
@@ -123,7 +117,7 @@ function throwIfInvalidManifest(manifest, pluginDir) {
   debug(result);
 
   if (!result.valid) {
-    const msgs = generateErrorMessages(result.errors);
+    const msgs = generateErrorMessages(result.errors!);
     console.error("Invalid manifest.json:");
     msgs.forEach((msg) => {
       console.error(`- ${msg}`);
@@ -134,34 +128,24 @@ function throwIfInvalidManifest(manifest, pluginDir) {
 
 /**
  * Create and save plugin.zip
- *
- * @param {string} outputPath
- * @param {!Buffer} plugin
- * @return {!Promise<string>} The value is output path of plugin.zip.
  */
-function outputPlugin(outputPath, plugin) {
+function outputPlugin(outputPath: string, plugin: Buffer): Promise<string> {
   return writeFile(outputPath, plugin).then((arg) => outputPath);
 }
 
 /**
  * Load JSON file without caching
- *
- * @param {string} jsonPath
- * @return {Object}
  */
-function loadJson(jsonPath) {
+function loadJson(jsonPath: string) {
   const content = fs.readFileSync(jsonPath, "utf8");
   return JSON.parse(content);
 }
 
 /**
  * Return validator for `relative-path` format
- *
- * @param {string} pluginDir
- * @return {function(string): boolean}
  */
-function validateRelativePath(pluginDir) {
-  return (str) => {
+function validateRelativePath(pluginDir: string) {
+  return (str: string) => {
     try {
       const stat = fs.statSync(path.join(pluginDir, str));
       return stat.isFile();
@@ -173,12 +157,9 @@ function validateRelativePath(pluginDir) {
 
 /**
  * Return validator for `maxFileSize` keyword
- *
- * @param {string} pluginDir
- * @return {function(number, string): boolean}
  */
-function validateMaxFileSize(pluginDir) {
-  return (maxBytes, filePath) => {
+function validateMaxFileSize(pluginDir: string) {
+  return (maxBytes: number, filePath: string) => {
     try {
       const stat = fs.statSync(path.join(pluginDir, filePath));
       return stat.size <= maxBytes;
