@@ -1,15 +1,10 @@
-"use strict";
-
 const N_TO_A = "a".charCodeAt(0) - "0".charCodeAt(0);
 const A_TO_K = "k".charCodeAt(0) - "a".charCodeAt(0);
 
 /**
  * `tr '0-9a-f' 'a-p'` in JS
- *
- * @param {string} hex like '8a7f7d'
- * @return {string}
  */
-function hex2a(hex) {
+export function hex2a(hex: string): string {
   return Array.from(hex)
     .map((s) => {
       if (s >= "0" && s <= "9") {
@@ -21,5 +16,3 @@ function hex2a(hex) {
     })
     .join("");
 }
-
-module.exports = hex2a;
