@@ -2,140 +2,141 @@ import { MockClient, buildMockClient } from "../../http/MockClient";
 import { AppClient } from "../AppClient";
 import { KintoneRequestConfigBuilder } from "../../KintoneRequestConfigBuilder";
 
+const APP_ID = 1;
+const REVISION = 5;
+const RECORD_ID = 3;
+const properties = {
+  fieldCode: {
+    type: "SINGLE_LINE_TEXT" as const,
+    code: "fieldCode",
+    label: "Text Field",
+  },
+};
+
+const layout = [
+  {
+    type: "ROW" as const,
+    fields: [
+      {
+        type: "SINGLE_LINE_TEXT",
+        code: "fieldCode1",
+        size: { width: "100" },
+      },
+      {
+        type: "LABEL",
+        label: "label1",
+        size: { width: "100" },
+      },
+      {
+        type: "SPACER",
+        elementId: "space",
+        size: { width: "100", height: "50" },
+      },
+    ],
+  },
+  {
+    type: "SUBTABLE" as const,
+    code: "tableFieldCode",
+    fields: [
+      {
+        type: "MULTI_LINE_TEXT",
+        code: "fieldCode2",
+        size: { width: "150", innerHeight: "200" },
+      },
+    ],
+  },
+  {
+    type: "GROUP" as const,
+    code: "fieldCode3",
+    layout: [
+      {
+        type: "ROW" as const,
+        fields: [
+          {
+            type: "NUMBER",
+            code: "fieldCode3_1",
+            size: {
+              width: 200,
+            },
+          },
+        ],
+      },
+    ],
+  },
+];
+
+const views = {
+  view1: {
+    type: "LIST" as const,
+    index: 0,
+    name: "view1",
+    fields: ["field"],
+    filterCond: 'field = "foo"',
+    sort: "sortField desc",
+  },
+  view2: {
+    type: "CALENDAR" as const,
+    index: 1,
+    name: "view2",
+    date: "dateField",
+    title: "titleField",
+    filterCond: 'field = "bar"',
+    sort: "sortField asc",
+  },
+  view3: {
+    type: "CUSTOM" as const,
+    index: 2,
+    name: "view3",
+    html: "<div>Hello!</div>",
+    pager: true,
+    device: "DESKTOP" as const,
+  },
+};
+
+const states = {
+  status1: {
+    name: "status1",
+    index: 0,
+    assignee: {
+      type: "ONE" as const,
+      entities: [
+        { entity: { type: "FIELD_ENTITY" as const, code: "creator" } },
+      ],
+    },
+  },
+  status2: {
+    name: "status2",
+    index: 1,
+    assignee: {
+      type: "ANY" as const,
+      entities: [{ entity: { type: "CREATOR" as const } }],
+    },
+  },
+  status3: {
+    name: "status3",
+    index: 2,
+    assignee: {
+      type: "ALL" as const,
+      entities: [
+        { entity: { type: "USER" as const, code: "user1" } },
+        { entity: { type: "USER" as const, code: "user2" } },
+      ],
+    },
+  },
+};
+const actions = [
+  { name: "action1to2", from: "status1", to: "status2" },
+  {
+    name: "action2to3",
+    from: "status2",
+    to: "status3",
+    filterCond: 'field = "foo"',
+  },
+];
+
 describe("AppClient", () => {
   let mockClient: MockClient;
   let appClient: AppClient;
-  const APP_ID = 1;
-  const REVISION = 5;
-  const RECORD_ID = 3;
-  const properties = {
-    fieldCode: {
-      type: "SINGLE_LINE_TEXT" as const,
-      code: "fieldCode",
-      label: "Text Field",
-    },
-  };
-
-  const layout = [
-    {
-      type: "ROW" as const,
-      fields: [
-        {
-          type: "SINGLE_LINE_TEXT",
-          code: "fieldCode1",
-          size: { width: "100" },
-        },
-        {
-          type: "LABEL",
-          label: "label1",
-          size: { width: "100" },
-        },
-        {
-          type: "SPACER",
-          elementId: "space",
-          size: { width: "100", height: "50" },
-        },
-      ],
-    },
-    {
-      type: "SUBTABLE" as const,
-      code: "tableFieldCode",
-      fields: [
-        {
-          type: "MULTI_LINE_TEXT",
-          code: "fieldCode2",
-          size: { width: "150", innerHeight: "200" },
-        },
-      ],
-    },
-    {
-      type: "GROUP" as const,
-      code: "fieldCode3",
-      layout: [
-        {
-          type: "ROW" as const,
-          fields: [
-            {
-              type: "NUMBER",
-              code: "fieldCode3_1",
-              size: {
-                width: 200,
-              },
-            },
-          ],
-        },
-      ],
-    },
-  ];
-
-  const views = {
-    view1: {
-      type: "LIST" as const,
-      index: 0,
-      name: "view1",
-      fields: ["field"],
-      filterCond: 'field = "foo"',
-      sort: "sortField desc",
-    },
-    view2: {
-      type: "CALENDAR" as const,
-      index: 1,
-      name: "view2",
-      date: "dateField",
-      title: "titleField",
-      filterCond: 'field = "bar"',
-      sort: "sortField asc",
-    },
-    view3: {
-      type: "CUSTOM" as const,
-      index: 2,
-      name: "view3",
-      html: "<div>Hello!</div>",
-      pager: true,
-      device: "DESKTOP" as const,
-    },
-  };
-
-  const states = {
-    status1: {
-      name: "status1",
-      index: 0,
-      assignee: {
-        type: "ONE" as const,
-        entities: [
-          { entity: { type: "FIELD_ENTITY" as const, code: "creator" } },
-        ],
-      },
-    },
-    status2: {
-      name: "status2",
-      index: 1,
-      assignee: {
-        type: "ANY" as const,
-        entities: [{ entity: { type: "CREATOR" as const } }],
-      },
-    },
-    status3: {
-      name: "status3",
-      index: 2,
-      assignee: {
-        type: "ALL" as const,
-        entities: [
-          { entity: { type: "USER" as const, code: "user1" } },
-          { entity: { type: "USER" as const, code: "user2" } },
-        ],
-      },
-    },
-  };
-  const actions = [
-    { name: "action1to2", from: "status1", to: "status2" },
-    {
-      name: "action2to3",
-      from: "status2",
-      to: "status3",
-      filterCond: 'field = "foo"',
-    },
-  ];
 
   beforeEach(() => {
     const requestConfigBuilder = new KintoneRequestConfigBuilder({
@@ -716,6 +717,44 @@ describe("AppClient", () => {
     });
   });
 
+  describe("getPerRecordNotifications", () => {
+    const params = {
+      app: APP_ID,
+    };
+    describe("without preview", () => {
+      beforeEach(async () => {
+        await appClient.getPerRecordNotifications(params);
+      });
+      it("should pass the path to the http client", () => {
+        expect(mockClient.getLogs()[0].path).toBe(
+          "/k/v1/app/notifications/perRecord.json"
+        );
+      });
+      it("should send a get request", () => {
+        expect(mockClient.getLogs()[0].method).toBe("get");
+      });
+      it("should pass app as a param to the http client", () => {
+        expect(mockClient.getLogs()[0].params).toEqual(params);
+      });
+    });
+    describe("preview: true", () => {
+      beforeEach(async () => {
+        await appClient.getPerRecordNotifications({ ...params, preview: true });
+      });
+      it("should pass the path to the http client", () => {
+        expect(mockClient.getLogs()[0].path).toBe(
+          "/k/v1/preview/app/notifications/perRecord.json"
+        );
+      });
+      it("should send a get request", () => {
+        expect(mockClient.getLogs()[0].method).toBe("get");
+      });
+      it("should pass app and preview as a param to the http client", () => {
+        expect(mockClient.getLogs()[0].params).toEqual(params);
+      });
+    });
+  });
+
   describe("getAppAcl", () => {
     const params = {
       app: APP_ID,
@@ -936,6 +975,8 @@ describe("AppClient", () => {
           ],
         },
       ],
+      notifyToCommenter: true,
+      revision: 1,
     };
 
     beforeEach(async () => {
@@ -953,11 +994,86 @@ describe("AppClient", () => {
       expect(mockClient.getLogs()[0].params).toEqual(params);
     });
   });
+
+  describe("getAppNotificationsReminder", () => {
+    const lang = "default";
+    const params = { app: APP_ID, lang } as const;
+    describe("without preview", () => {
+      beforeEach(async () => {
+        await appClient.getReminderNotifications(params);
+      });
+      it("should pass the path to the http client", () => {
+        expect(mockClient.getLogs()[0].path).toBe(
+          "/k/v1/app/notifications/reminder.json"
+        );
+      });
+      it("should send a get request", () => {
+        expect(mockClient.getLogs()[0].method).toBe("get");
+      });
+      it("should pass app and lang as a param to the http client", () => {
+        expect(mockClient.getLogs()[0].params).toEqual(params);
+      });
+    });
+    describe("preview: true", () => {
+      beforeEach(async () => {
+        await appClient.getReminderNotifications({
+          ...params,
+          preview: true,
+        });
+      });
+      it("should pass the path to the http client", () => {
+        expect(mockClient.getLogs()[0].path).toBe(
+          "/k/v1/preview/app/notifications/reminder.json"
+        );
+      });
+      it("should send a get request", () => {
+        expect(mockClient.getLogs()[0].method).toBe("get");
+      });
+      it("should pass app and lang as a param to the http client", () => {
+        expect(mockClient.getLogs()[0].params).toEqual(params);
+      });
+    });
+  });
+  describe("updateAppUpdateGeneralNotifications", () => {
+    const params = {
+      app: APP_ID,
+      notifications: [
+        {
+          entity: {
+            type: "USER" as const,
+            code: "foo",
+          },
+          includeSubs: true,
+          recordAdded: true,
+          recordEdited: true,
+          commentAdded: true,
+          statusChanged: true,
+          fileImported: true,
+          notifyToCommenter: true,
+        },
+      ],
+      notifyToCommenter: true,
+      revision: 1,
+    };
+    beforeEach(async () => {
+      await appClient.updateGeneralNotifications(params);
+    });
+    it("should pass the path to the http client", () => {
+      expect(mockClient.getLogs()[0].path).toBe(
+        "/k/v1/preview/app/notifications/general.json"
+      );
+    });
+    it("should send a put request", () => {
+      expect(mockClient.getLogs()[0].method).toBe("put");
+    });
+    it("should pass app and rights as a param to the http client", () => {
+      expect(mockClient.getLogs()[0].params).toEqual(params);
+    });
+  });
 });
 
 describe("AppClient with guestSpaceId", () => {
   it("should pass the path to the http client", async () => {
-    const APP_ID = 1;
     const GUEST_SPACE_ID = 2;
     const lang = "default";
     const params = { app: APP_ID, lang } as const;
