@@ -1,6 +1,22 @@
 import { Entity } from "../entity";
 
-export type GeneralNotificationsForResponse = {
+export type GeneralNotificationForParameter = {
+  entity:
+    | Entity
+    | {
+        type: "FIELD_ENTITY";
+        code: string;
+      };
+  includeSubs?: boolean;
+  recordAdded?: boolean;
+  recordEdited?: boolean;
+  commentAdded?: boolean;
+  statusChanged?: boolean;
+  fileImported?: boolean;
+  notifyToCommenter?: boolean;
+};
+
+export type GeneralNotificationForResponse = {
   entity:
     | Entity
     | {
@@ -15,11 +31,38 @@ export type GeneralNotificationsForResponse = {
   fileImported: boolean;
 };
 
-export type PerRecordNotificationsForParameter = {
+export type PerRecordNotificationForParameter = {
   filterCond: string;
   title?: string;
   targets: Array<{
     entity: Entity | { type: "FIELD_ENTITY"; code: string };
+    includeSubs?: boolean;
+  }>;
+};
+
+export type PerRecordNotificationForResponse = {
+  filterCond: string;
+  title: string;
+  targets: Array<{
+    entity: Entity | { type: "FIELD_ENTITY"; code: string };
     includeSubs: boolean;
   }>;
+  revision: string;
+};
+
+export type ReminderNotificationForResponse = {
+  timing: {
+    code: string;
+    daysLater: string;
+    hoursLater?: string;
+    time?: string;
+  };
+  filterCond: string;
+  title: string;
+  targets: [
+    {
+      entity: Entity | { type: "FIELD_ENTITY"; code: string };
+      includeSubs: boolean;
+    }
+  ];
 };
